+++ src/main/java/ru/yandex/praktikum/client/OrdersClient.java
@@ -1,5 +1,6 @@
 package ru.yandex.praktikum.client;
 
+import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 import ru.yandex.praktikum.pojo.CreateOrderRequest;
 
@@ -9,6 +10,7 @@ public class OrdersClient extends RestClient {
     private static final String CREATE_AN_ORDER = "/api/orders";
     private static final String GET_ORDERS = "/api/orders";
 
+    @Step("Create Order Without Authorization Request")
     public ValidatableResponse createOrderWithoutAuthorization(CreateOrderRequest createOrderRequest) {
         return given()
                 .spec(getDefaultRequestSpec())
@@ -17,6 +19,7 @@ public class OrdersClient extends RestClient {
                 .then();
     }
 
+    @Step("Create Order With Authorization Request")
     public ValidatableResponse createOrderWithAuthorization(CreateOrderRequest createOrderRequest, String accessToken) {
         return given()
                 .spec(getDefaultRequestSpec())
@@ -26,6 +29,7 @@ public class OrdersClient extends RestClient {
                 .then();
     }
 
+    @Step("Get Orders List With Authorization Request")
     public ValidatableResponse getOrdersListWithAuthorization(String accessToken) {
         return given()
                 .spec(getDefaultRequestSpec())
@@ -34,6 +38,7 @@ public class OrdersClient extends RestClient {
                 .then();
     }
 
+    @Step("Get Order's List Without Authorization Request Negative")
     public ValidatableResponse getOrdersListWithoutAuthorizationNEGATIVE() {
         return given()
                 .spec(getDefaultRequestSpec())
