+++ src/test/java/ru/yandex/praktikum/user/LoginUserTest.java
@@ -1,6 +1,5 @@
 package ru.yandex.praktikum.user;
 
-import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.http.HttpStatus;
@@ -8,6 +7,7 @@ import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.jupiter.api.DisplayName;
 import ru.yandex.praktikum.TestSteps;
 import ru.yandex.praktikum.client.UserClient;
 import ru.yandex.praktikum.pojo.CreateUserRequest;
@@ -40,7 +40,7 @@ public class LoginUserTest {
         CreateUserRequest createUserRequest = getRandomCreateUserRequest();
         ValidatableResponse createUserResponse = testSteps.createUserAndReturnResponse(createUserRequest);
         LoginUserRequest loginUserRequest = getLoginUserRequest(createUserRequest.getEmail(), createUserRequest.getPassword());
-        ValidatableResponse loginUserResponse = testSteps.loginAndReturnResponse(loginUserRequest);
+        testSteps.loginAndReturnResponse(loginUserRequest);
         accessToken = testSteps.getAnAccessTokenFromResponse(createUserResponse);
     }
